Initial Commit
The Notepad in Python is a simple and similar version of the text editor application. In this application, you can create text documents easily. The system allows you greater flexibility while editing the text files or the documents. Talking about this system, this system has various features intact within it.